pyproject.toml
[tool.poetry]
name = "practice"
version = "0.1.0"
description = ""
authors = ["UASTRONGMAN <shkolniy.maks@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
